+++ src/GraphicLibs/ncurses/Ncurses.hpp
@@ -16,7 +16,7 @@ namespace Ncurses {
             DisplayModule();
             ~DisplayModule();
 
-            void update(float delta,
+            void update(double delta,
             Arcade::ECS::IEventManager &eventManager,
             Arcade::ECS::IEntityManager &entityManager) final;
