+++ src/GraphicLibs/ncurses/SPRITE.hpp
@@ -18,7 +18,7 @@ namespace Ncurses {
 
         class SpriteSystem : public Arcade::ECS::ISystem {
             public:
-                void run(float deltaTime,
+                void run(double deltaTime,
                 Arcade::ECS::IEventManager &eventManager,
                 Arcade::ECS::IEntityManager &entityManager) final;
